iot-controller/requirements.txt
fastapi
uvicorn
pika
pymongo
prometheus_client
python-logstash~=0.4.8